# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ФИТИТ — Персональный план питания API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /survey/basic:
    post:
      operationId: submitBasicSurvey
      tags: [survey]
      summary: Submit basic survey (survey 1) — gender, goal, hormones, lifestyle, measurements, activity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BasicSurveyInput"
      responses:
        "200":
          description: Metabolic result calculated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetabolicResult"
  /survey/preferences:
    post:
      operationId: submitPreferencesSurvey
      tags: [survey]
      summary: Submit preferences survey (survey 2) — eating habits, allergies, health, budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferencesSurveyInput"
      responses:
        "200":
          description: Preferences saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /survey/result:
    get:
      operationId: getMetabolicResult
      tags: [survey]
      summary: Get the current user's latest metabolic result
      responses:
        "200":
          description: Metabolic result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MetabolicResult"
        "404":
          description: No result yet
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    BasicSurveyInput:
      type: object
      required: [gender, goal, hormonalDisorder, lifestyle, age, height, weight, targetWeight, dailySteps, cardioMinutesPerWeek, strengthMinutesPerWeek]
      properties:
        gender:
          type: string
          enum: [male, female]
        goal:
          type: string
          enum: [lose_weight, maintain_recompose, gain_weight]
        hormonalDisorder:
          type: string
          enum: [none, hypothyroidism, insulin_leptin_resistance, sex_hormone_deficiency, other_endocrine]
        lifestyle:
          type: string
          enum: [sedentary, office_active, on_feet, intense_training, physical_labor]
        age:
          type: integer
        height:
          type: integer
        weight:
          type: number
        targetWeight:
          type: number
        sport:
          type: ["string", "null"]
        dailySteps:
          type: integer
        cardioMinutesPerWeek:
          type: integer
        strengthMinutesPerWeek:
          type: integer
        waistCm:
          type: ["number", "null"]
        weightChangePast12Months:
          type: ["string", "null"]
        energyMorning:
          type: ["integer", "null"]
        energyAfternoon:
          type: ["integer", "null"]
        energyEvening:
          type: ["integer", "null"]
        sleepFallingAsleep:
          type: ["string", "null"]
        sleepNightWakeups:
          type: ["string", "null"]
        sleepRestfulness:
          type: ["string", "null"]
        chronicStressLevel:
          type: ["integer", "null"]
        caffeineIntake:
          type: ["string", "null"]
        bloatingAfterEating:
          type: ["string", "null"]
        gasForming:
          type: ["string", "null"]
        heavinessAfterEating:
          type: ["string", "null"]
        stoolRegularity:
          type: ["string", "null"]
        foodReactions:
          type: array
          items:
            type: string
        skinHairNailSymptoms:
          type: array
          items:
            type: string
        illnessFrequency:
          type: ["string", "null"]
        recoverySpeed:
          type: ["string", "null"]
        chronicInflammations:
          type: array
          items:
            type: string
        allergies:
          type: ["string", "null"]
        generalSymptoms:
          type: array
          items:
            type: string
        menstrualRegularity:
          type: ["string", "null"]
        pmsLevel:
          type: ["string", "null"]
        menstrualPain:
          type: ["string", "null"]
        edemaLevel:
          type: ["string", "null"]
        morningErections:
          type: ["string", "null"]
        trainingRecovery:
          type: ["string", "null"]
        proteinFrequency:
          type: ["string", "null"]
        fishFrequency:
          type: ["string", "null"]
        vegetableFrequency:
          type: ["string", "null"]
        fruitFrequency:
          type: ["string", "null"]
        wholeGrainFrequency:
          type: ["string", "null"]
        carbType:
          type: ["string", "null"]
    MetabolicResult:
      type: object
      required: [bmr, tdee, targetCalories, proteins, fats, carbs, bmi, bmiCategory, weightToGoal, totalDeficitNeeded, estimatedMonths, progressBlockers]
      properties:
        id:
          type: ["integer", "null"]
        bmr:
          type: integer
        tdee:
          type: integer
        targetCalories:
          type: integer
        proteins:
          type: integer
        fats:
          type: integer
        carbs:
          type: integer
        bmi:
          type: number
        bmiCategory:
          type: string
        weightToGoal:
          type: number
        totalDeficitNeeded:
          type: integer
        estimatedMonths:
          type: integer
        progressBlockers:
          type: array
          items:
            type: string
        moderateDeficit:
          type: ["integer", "null"]
        aggressiveDeficit:
          type: ["integer", "null"]
    PreferencesSurveyInput:
      type: object
      required: [mealsPerDay, dietType]
      properties:
        mealsPerDay:
          type: string
          enum: [one_two, three, four_five, six_plus]
        dietType:
          type: string
          enum: [omnivore, vegetarian, vegan, pescatarian, gluten_free, keto]
        dislikedFoods:
          type: ["string", "null"]
        foodAllergyOrIntolerance:
          type: ["string", "null"]
        favoriteFoods:
          type: ["string", "null"]
        culturalRestrictions:
          type: ["string", "null"]
        chronicDiseases:
          type: ["string", "null"]
        medications:
          type: ["string", "null"]
        physicalLimitations:
          type: ["string", "null"]
        stressLevel:
          type: ["string", "null"]
          enum: [low, medium, high, null]
        previousWeightAttempts:
          type: ["string", "null"]
        previousDiets:
          type: ["string", "null"]
        eatingDisorders:
          type: ["string", "null"]
        snacksPerDay:
          type: ["integer", "null"]
        favoriteSnacks:
          type: ["string", "null"]
        sweetFrequency:
          type: ["string", "null"]
          enum: [daily, several_per_week, rarely, never, null]
        sugarAddiction:
          type: ["boolean", "null"]
        budget:
          type: ["string", "null"]
          enum: [minimal, medium, high, null]
        cookingTimeMinutes:
          type: ["string", "null"]
          enum: [under_15, fifteen_to_30, over_30, null]
        coffeeTeaCupsPerDay:
          type: ["string", "null"]
          enum: [zero, one_two, three_four, five_plus, null]
        waterLitersPerDay:
          type: ["string", "null"]
          enum: [zero, half, one_two, two_three, three_plus, null]
        motivationLevel:
          type: ["string", "null"]
          enum: [under_50, fifty_to_75, over_75, null]
        motivationHelpers:
          type: array
          items:
            type: string
...