# lib/api-zod/src/generated/types/preferencesSurveyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ФИТИТ — Персональный план питания API
 * OpenAPI spec version: 0.1.0
 */
import type { PreferencesSurveyInputBudget } from './preferencesSurveyInputBudget';
import type { PreferencesSurveyInputCoffeeTeaCupsPerDay } from './preferencesSurveyInputCoffeeTeaCupsPerDay';
import type { PreferencesSurveyInputCookingTimeMinutes } from './preferencesSurveyInputCookingTimeMinutes';
import type { PreferencesSurveyInputDietType } from './preferencesSurveyInputDietType';
import type { PreferencesSurveyInputMealsPerDay } from './preferencesSurveyInputMealsPerDay';
import type { PreferencesSurveyInputMotivationLevel } from './preferencesSurveyInputMotivationLevel';
import type { PreferencesSurveyInputStressLevel } from './preferencesSurveyInputStressLevel';
import type { PreferencesSurveyInputSweetFrequency } from './preferencesSurveyInputSweetFrequency';
import type { PreferencesSurveyInputWaterLitersPerDay } from './preferencesSurveyInputWaterLitersPerDay';

export interface PreferencesSurveyInput {
  mealsPerDay: PreferencesSurveyInputMealsPerDay;
  dietType: PreferencesSurveyInputDietType;
  /** @nullable */
  dislikedFoods?: string | null;
  /** @nullable */
  foodAllergyOrIntolerance?: string | null;
  /** @nullable */
  favoriteFoods?: string | null;
  /** @nullable */
  culturalRestrictions?: string | null;
  /** @nullable */
  chronicDiseases?: string | null;
  /** @nullable */
  medications?: string | null;
  /** @nullable */
  physicalLimitations?: string | null;
  /** @nullable */
  stressLevel?: PreferencesSurveyInputStressLevel;
  /** @nullable */
  previousWeightAttempts?: string | null;
  /** @nullable */
  previousDiets?: string | null;
  /** @nullable */
  eatingDisorders?: string | null;
  /** @nullable */
  snacksPerDay?: number | null;
  /** @nullable */
  favoriteSnacks?: string | null;
  /** @nullable */
  sweetFrequency?: PreferencesSurveyInputSweetFrequency;
  /** @nullable */
  sugarAddiction?: boolean | null;
  /** @nullable */
  budget?: PreferencesSurveyInputBudget;
  /** @nullable */
  cookingTimeMinutes?: PreferencesSurveyInputCookingTimeMinutes;
  /** @nullable */
  coffeeTeaCupsPerDay?: PreferencesSurveyInputCoffeeTeaCupsPerDay;
  /** @nullable */
  waterLitersPerDay?: PreferencesSurveyInputWaterLitersPerDay;
  /** @nullable */
  motivationLevel?: PreferencesSurveyInputMotivationLevel;
  motivationHelpers?: string[];
}